Fix CI problem. (#10442)
* Fix CI problem.

* Fix CI problem.

* Fix CI problem.

* Fix CI problem.
