+++ test/naming-test/src/test/java/com/alibaba/nacos/test/naming/NamingTlsServiceTls_ITCase.java
@@ -26,6 +26,7 @@ import com.alibaba.nacos.common.remote.client.RpcConstants;
 import com.alibaba.nacos.core.remote.RpcServerTlsConfig;
 import org.junit.Assert;
 import org.junit.FixMethodOrder;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.MethodSorters;
@@ -54,6 +55,7 @@ import static com.alibaba.nacos.test.naming.NamingBase.randomDomainName;
         RpcServerTlsConfig.PREFIX+".certPrivateKey=test-server-key.pem",
 },
         webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@Ignore("TODO, Fix cert expired problem")
 public class NamingTlsServiceTls_ITCase {
